Add attack for bat (shooter and bullet)

+++ include/headers/transform_component.h
@@ -3,18 +3,18 @@
 class TransformComponent
 {
 public:
-    int x, y;
-    int speed;
+    double x, y;
+    double speed;
 
     // Velocity (Direction) of x and y
-    int vx, vy;
+    double vx, vy;
 
     int h, w;
-    TransformComponent(int _x, int _y, int _speed, int _w, int _h);
+    TransformComponent(int _x, int _y, double _speed, int _w, int _h);
     ~TransformComponent();
     void Update();
-    void SetDir(int _vx, int _vy);
+    void SetDir(double _vx, double _vy);
     void SetPosition(int _x, int _y);
-    void SetSpeed(int _speed);
-    void AddSpeed(int v);
+    void SetSpeed(double _speed);
+    void AddSpeed(double v);
 };
