+++ src/transform_component.cpp
@@ -1,6 +1,6 @@
 #include <transform_component.h>
 
-TransformComponent::TransformComponent(int _x, int _y, int _speed, int _w, int _h)
+TransformComponent::TransformComponent(int _x, int _y, double _speed, int _w, int _h)
 {
     x = _x;
     y = _y;
@@ -20,7 +20,7 @@ void TransformComponent::Update()
     x += speed * vx;
     y += speed * vy;
 }
-void TransformComponent::SetDir(int _vx, int _vy)
+void TransformComponent::SetDir(double _vx, double _vy)
 {
     vx = _vx;
     vy = _vy;
@@ -30,11 +30,11 @@ void TransformComponent::SetPosition(int _x, int _y)
     x = _x;
     y = _y;
 }
-void TransformComponent::SetSpeed(int _speed)
+void TransformComponent::SetSpeed(double _speed)
 {
     speed = _speed;
 }
-void TransformComponent::AddSpeed(int v)
+void TransformComponent::AddSpeed(double v)
 {
     speed += v;
 }
